requirements.txt
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
filelock==3.20.0
fsspec==2025.12.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
joblib==1.5.2
morfeusz2==1.99.12
networkx==3.6
nltk==3.9.2
numpy==2.3.5
packaging==25.0
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
soupsieve==2.8
tokenizers==0.22.1
tqdm==4.67.1
transformers==4.57.3
typing_extensions==4.15.0
urllib3==2.6.0
playwright
readability-lxml
sentence-transformers
